Move atomic_number back into atomic.py
