+++ plasmapy/atomic/__init__.py
@@ -1,21 +1,28 @@
 """Physical and mathematical constants for use within PlasmaPy."""
 
-from .names import (atomic_symbol,
-                    isotope_symbol,
-                    element_name,
-                    atomic_number)
+from .names import (
+    atomic_symbol,
+    isotope_symbol,
+    element_name,
+)
 
-from .atomic import (is_isotope_stable,
-                     half_life,
-                     mass_number,
-                     standard_atomic_weight,
-                     isotope_mass,
-                     ion_mass,
-                     known_isotopes,
-                     common_isotopes,
-                     stable_isotopes,
-                     isotopic_abundance,
-                     charge_state,
-                     electric_charge)
+from .atomic import (
+    atomic_number,
+    is_isotope_stable,
+    half_life,
+    mass_number,
+    standard_atomic_weight,
+    isotope_mass,
+    ion_mass,
+    known_isotopes,
+    common_isotopes,
+    stable_isotopes,
+    isotopic_abundance,
+    charge_state,
+    electric_charge,
+)
 
-from .nuclear import nuclear_binding_energy, nuclear_reaction_energy
+from .nuclear import (
+    nuclear_binding_energy,
+    nuclear_reaction_energy,
+)
